==> Mouse_E18_Brain/notebooks/MouseEmbryonicBrain/1C_oligo_signal_noise.Rmd <==
---
title: "Signal to noise plots"
author: "Daniel Brown"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      smooth_scroll: true
    theme: readable
    highlight: tango 
    df_print: paged
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      dev = "png",
                      dpi = 300,
                      fig.asp=0.6,
                      fig.path='./figures/signalNoise/',
                      layout="l-body-outset")
knitr::opts_knit$set(root.dir = here::here())
library(Matrix)
library(tidyverse)
library(FSA)
library(ggridges)
library(DropletUtils)
library(patchwork)
source(here::here("scripts/R/theme_publication.R"))
theme_set(theme_Publication())
```

## Aim

Rough analysis to visualise data and decide which plots to use for manuscript. I will make a new notebook containing a subset of useful figures to polish up for presentation in talks and manuscripts.

## Recap

Run a more challenging sample = Mouse embryonic brain E18.5
We took mouse #3, the largest one for this purpose

## Protocols

1. MULTI-Seq LMO, LMO = Capture 1
2. 10x Genomics CellPlex, CMO = Capture 2
3. Custom Nextera CMO, Nxt = Capture 3

# Read data

These files are generated in scripts/R/NN236_write_tagCounts_file.R

```{r loadData}
CMO_matrix_dir <- here::here(
  "Mouse_E18_Brain/data/Matrices", "cmo_tag_counts")

LMO_matrix_dir = here::here(
  "Mouse_E18_Brain/data/Matrices", "lmo_tag_counts")

HTO_matrix_dir = here::here(
  "Mouse_E18_Brain/data/Matrices", "nxt_tag_counts")

matrix_dir <- list(CMO_matrix_dir, LMO_matrix_dir, HTO_matrix_dir)
names(matrix_dir) <- c("CMO", "LMO", "Nxt")
```

# Read into tibble format {.tabset}

## CellPlex

```{r readCMO}
mat <- DropletUtils::read10xCounts(matrix_dir[["CMO"]])

cm <- as.data.frame(
  Matrix::as.matrix(
  t(counts(mat)
  )))
  
tb <- cm %>% 
  pivot_longer(cols = c(CMO301:CMO312),
               names_to = "Multiplexing_Capture", 
               values_to = "Count"
              )
tb$Protocol <- "CMO"
# Need to add 1 before log
tb$Count <- tb$Count + 1
tb_cmo <- tb
```

## MULTI-Seq LMO

```{r readLMO}
mat <-  DropletUtils::read10xCounts(matrix_dir[["LMO"]])
cm <- as.data.frame(
  Matrix::as.matrix(
  t(counts(mat)
  )))
  
tb <- cm %>% 
  pivot_longer(cols = c(MULTI_2:MULTI_13),
               names_to = "Multiplexing_Capture", 
               values_to = "Count"
              )
tb$Protocol <- "LMO"
# Need to add 1 before log
tb$Count <- tb$Count + 1

tb_lmo <- tb
```

## MULTI-Seq Nxt

```{r readNxt}
mat <-  DropletUtils::read10xCounts(matrix_dir[["Nxt"]])
cm <- as.data.frame(
  Matrix::as.matrix(
  t(counts(mat)
  )))
  
tb <- cm %>% 
  pivot_longer(cols = c(Nxt_451:Nxt_465),
               names_to = "Multiplexing_Capture", 
               values_to = "Count"
              )
tb$Protocol <- "Nxt"
# Need to add 1 before log
tb$Count <- tb$Count + 1

tb_Nxt <- tb
```

We end by saving objects

```{r save}
tb <- list(
  tb_cmo,
  tb_lmo,
  tb_Nxt
)

names(tb) <- c("CMO", "LMO" ,"Nxt")
saveRDS(tb, here::here(
  "Mouse_E18_Brain/data/Matrices", "signal_noise_counts.tb.rds"
))
```

Restart from saved matrix

```{r}
tb <- readRDS(here::here(
  "Mouse_E18_Brain/data/Matrices/signal_noise_counts.tb.rds"
))
```


# GGridges {.tabset}

The plot I liked most is [https://cran.r-project.org/web/packages/ggridges/vignettes/introduction.html](GGridges)

## CMO

CellPlex diluted 1:10 had the best signal to noise.

```{r plt_cmo, fig.cap="CellPlex signal to noise"}
plt_cmo <- ggplot(tb_cmo, aes(x = Count, y = Multiplexing_Capture, 
                   fill = Multiplexing_Capture)) +
  geom_density_ridges() +
  scale_x_continuous(trans='log10', limits = c(1,1e5)) +
  annotation_logticks(base = 10, sides = "b") +
  theme_ridges(font_size = 18) + 
  #scale_fill_brewer(palette = "Set1") +
  ylab("") + xlab("") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  theme(legend.position = "none")

plt_cmo
```

## LMO

MULTI-Seq did not work at all. Should have washed out the media properly.

```{r plt_lmo, fig.cap="MULTI-Seq signal to noise"}
plt_lmo <- ggplot(tb_lmo, aes(x = Count, y = Multiplexing_Capture, 
                   fill = Multiplexing_Capture)) +
  geom_density_ridges() +
  scale_x_continuous(trans='log10', limits = c(1,1e5)) +
  annotation_logticks(base = 10, sides = "b") +
  theme_ridges(font_size = 18) + 
  #scale_fill_brewer(palette = "Dark2") +
  ylab("") + xlab("") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  theme(legend.position = "none")

plt_lmo
```

## Nxt

Trimodal peak is concerning and needs further investigation. Could be cell size or doublet related.

```{r plt_Nxt, fig.cap="Custom MULTI-Seq signal to noise"}
plt_Nxt <- ggplot(tb_Nxt, aes(x = Count, y = Multiplexing_Capture, 
                   fill = Multiplexing_Capture)) +
  geom_density_ridges() +
  scale_x_continuous(trans='log10', limits = c(1,1e5)) +
  annotation_logticks(base = 10, sides = "b") +
  theme_ridges(font_size = 18) + 
  #scale_fill_brewer(palette = "Dark2") +
  ylab("") + xlab("") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  theme(legend.position = "none")

plt_Nxt
```

# Gather ggridges into 1 plot

The aim is to reproduce the output of the cellranger report

```{r, fig.height=30, fig.width=8, fig.cap="Summary plot"}
plt_cmo / plt_lmo / plt_Nxt
```

# Plot signal to noise for each protocol

```{r}
head_path <- "/stornext/Projects/GenomicsRnD/brown.d/DB_NN258/multi_out/"
tail_path <- "/outs/per_sample_outs/Lox-brain-1/metrics_summary.csv"

CMO_file_path <- paste(head_path, "R010_CMO_ME", tail_path, sep="")
LMO_file_path <- paste(head_path, "R010_LMO_ME", tail_path, sep="")
Nxt_file_path <- paste(head_path, "R010_NXT_ME", tail_path, sep="")

Nxt_data = read_csv(Nxt_file_path)
Nxt_data <- Nxt_data %>% 
  filter(`Metric Name` == "CMO signal-to-noise ratio")

CMO_data = read_csv(CMO_file_path)
CMO_data <- CMO_data %>% 
  filter(`Metric Name` == "CMO signal-to-noise ratio")

LMO_data = read_csv(LMO_file_path)
LMO_data <- LMO_data %>% 
  filter(`Metric Name` == "CMO signal-to-noise ratio")

Nxt_data$Protocol <- "Custom MULTI-Seq CMO"
LMO_data$Protocol <- "MULTI-Seq LMO"
CMO_data$Protocol <- "CellPlex 1:10"

tb <- rbind(CMO_data, LMO_data, Nxt_data)
tb$`Metric Value` <- as.numeric(tb$`Metric Value`)
```

Recolour the tags to reflect the ggridges plot.
Mung the names of the tags

```{r}
tb$CellRangerCall <- tb$`Group Name`

tb$CellRangerCall <- recode(tb$CellRangerCall,
                      CMO301="Tag1",CMO302="Tag2",CMO303="Tag3",CMO304="Tag4",
                      CMO305="Tag5",CMO306="Tag6",CMO307="Tag7",CMO308="Tag8",
                      CMO309="Tag9",CMO310="Tag10",CMO311="Tag11",CMO312="Tag12",
                      Nxt_451="Tag1",Nxt_452="Tag2",Nxt_453= "Tag3",Nxt_455= "Tag4",
                      Nxt_456= "Tag5",Nxt_457= "Tag6",Nxt_458= "Tag7",Nxt_459= "Tag8",
                      Nxt_460= "Tag9",Nxt_462= "Tag10",Nxt_463= "Tag11",Nxt_465= "Tag12",
                      MULTI_2= "Tag1",MULTI_3= "Tag2",MULTI_4= "Tag3",MULTI_5= "Tag4",
                      MULTI_6= "Tag5", MULTI_7= "Tag6",MULTI_8= "Tag7",MULTI_9= "Tag8",
                      MULTI_10= "Tag9", MULTI_11= "Tag10",MULTI_12= "Tag11",MULTI_13= "Tag12"
)

tb$Protocol <- recode(tb$Protocol,
                      "Custom MULTI-Seq CMO"="Custom\nMULTI-Seq CMO")

write_csv(tb, here::here(
  "Mouse_E18_Brain/data/Matrices/signal_noise.csv"
))
```

## Make the plot

The signal to noise is above the recomended 5 for CellPlex diluted 1:10.  
Suprisingly based on the raw count data is the Nxt and LMO which looks different, the LMO and Nxt have similar signal to noise.

```{r signal_noise, fig.cap="Signal to noise from cell ranger"}
plt4 <- ggplot(data=tb) +
  geom_jitter(aes(y=`Metric Value`, x=Protocol, colour=CellRangerCall), 
              size=3, width = 0.15,alpha=0.66, show.legend=FALSE) +
  scale_y_continuous(breaks=seq(0,10,1)) +
  xlab("") + ylab("Signal to noise ratio") +
  scale_colour_brewer(type="qualitative", palette = "Dark2") +
  theme_Publication(base_size=16)

plt4
```

## Conduct a statistical test comparing the ratios

Data is not normally distributed

```{r}
# Basic Q-Q plot
qqnorm(tb$`Metric Value`)
qqline(tb$`Metric Value`, col = "red")  # Adds a reference line in red
```

Performing the Kruskal-Wallis test

```{r}
kruskal_result <- kruskal.test(`Metric Value` ~ Protocol, tb)

# View the results
print(kruskal_result)

```

After finding a significant result in the Kruskal-Wallis test with three groups, you'll want to perform post-hoc pairwise comparisons to determine which specific groups differ from each other. The Dunn test is commonly used for this purpose. In R, you can use the dunn.test package to perform Dunn's test.

```{r}
dunnTest(`Metric Value` ~ Protocol, tb,
         method="bonferroni")
```


## Session info
```{r sessionInfo}
sessionInfo()
```
