==> Mouse_E18_Brain/notebooks/MouseEmbryonicBrain/2C_GEX_cluster_visualise.Rmd <==
---
title: "GEX Clustering and dimension reduction"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      smooth_scroll: true
    theme: readable
    highlight: tango 
    df_print: paged
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      dev = "png",
                      dpi = 300,
                      fig.asp=0.6, 
                      fig.path='./figures/gex_clustering/',
                      warning=FALSE, message=FALSE)
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = here::here())
suppressPackageStartupMessages({
  library(tidyverse)
  library(patchwork)
  library(scuttle)
  library(scater)
  library(scran)
  })

source(here::here("scripts/R/theme_publication.R"))
theme_set(theme_Publication(base_size = 14))
```

# Aim

Visulaise the dimension reduction results with respect to the metadata.

## Samples

# Load Data

From the 2B_QC_dimred.Rmd notebook

```{r}
sce_list <- readRDS("/vast/scratch/users/brown.d/S000322/S000322.dimred.rds")
names(sce_list)

capture <- names(sce_list)
```

# Clustering

We build a shared nearest neighbour graph and use the Louvain algorithm to identify clusters. We build the graph using the principal components.

## Initial clustering

Smaller 'k' usually yields finer clusters.

Clustering is stable from k=15 to k=20 with 10-12 clusters resolved.  
The extra clusters from 10 to 12 are very few cells.

```{r cluster}
set.seed(911)

for (n in names(sce_list)) {
  current <- sce_list[[n]]
  snn_gr_cmo <- buildSNNGraph(current, use.dimred = "PCA", k = 20) 
  clusters <- igraph::cluster_louvain(snn_gr_cmo)
  current$cluster <- factor(clusters$membership)
  sce_list[[n]] <- current
}
```

# Visualise UMAP by QC parameters {.tabset}

## Cluster number

In CellPlex sample cluster 12 and 13 look like low quality cells.

* Low library size
* Higher mitochondrial percentage

However this could also reflect biology as all the cells in this analysis have passed median absolute deviation threshold of 2.

In the LMO sample all the cells on the left hand side have low library size and higher mito percentage. This makes sense as this is the most compromised sample.

I could say the same thing about the MULTI-Seq CMO clusters in the bottom left

```{r umap_cluster, fig.dim = c(20, 5)}
umap_plt <- list()

for (i in 1:length(sce_list)) {
  current <- sce_list[[i]]
  umap_plt[[i]] <- plotReducedDim(current, dimred="UMAP", colour_by="cluster") + 
    ggtitle(capture[i]) + theme_Publication()
}

umap_plt[[1]] + umap_plt[[2]] + umap_plt[[3]]
```

## Library size

```{r umap_libSize, fig.dim = c(20, 5)}
umap_plt <- list()

for (i in 1:length(sce_list)) {
  current <- sce_list[[i]]
  current$logSum <- log10(current$sum +1)
  umap_plt[[i]] <- plotReducedDim(current, dimred="UMAP", colour_by="logSum") + 
    ggtitle(capture[i]) + theme_Publication()
}

umap_plt[[1]] + umap_plt[[2]] + umap_plt[[3]]
```

## Mitochondrial percentage

```{r umap_mito, fig.dim = c(20, 5)}
umap_plt <- list()

for (i in 1:length(sce_list)) {
  current <- sce_list[[i]]
  umap_plt[[i]] <- plotReducedDim(current, dimred="UMAP", colour_by="subsets_Mito_percent") + 
    ggtitle(capture[i]) + theme_Publication()
}

umap_plt[[1]] + umap_plt[[2]] + umap_plt[[3]]
```

## Cell cycle

```{r umap_cc, fig.dim = c(20, 5)}
umap_plt <- list()

for (i in 1:length(sce_list)) {
  current <- sce_list[[i]]
  umap_plt[[i]] <- plotReducedDim(current, dimred="UMAP", colour_by="Mki67") + 
    ggtitle(capture[i]) + theme_Publication()
}

umap_plt[[1]] + umap_plt[[2]] + umap_plt[[3]]
```

# Visualise UMAP by multiplexing tag calls {.tabset}

## Cell Ranger multi

```{r cellranger_multi, fig.dim = c(20, 5)}
sce_tag <- sce_list[c(1:3)]

umap_plt <- list()

for (i in 1:length(sce_tag)) {
  current <- sce_tag[[i]]
  umap_plt[[i]] <- plotReducedDim(current, dimred="UMAP", colour_by="mul_Assignment") + 
  ggtitle(capture[i]) +
  theme_Publication()
}

umap_plt[[1]] + umap_plt[[2]] + umap_plt[[3]]
```

## BFF cluster

```{r bff_cluster, fig.dim = c(20, 5)}
umap_plt <- list()

for (i in 1:length(sce_tag)) {
  current <- sce_tag[[i]]
  umap_plt[[i]] <- plotReducedDim(current, dimred="UMAP", colour_by="bff_cluster") + 
  ggtitle(capture[i]) +
  theme_Publication()
}

umap_plt[[1]] + umap_plt[[2]] + umap_plt[[3]]
```

# Save SCE object

```{r saveSCEs, eval=FALSE}
saveRDS(sce_list, here::here(
  "Mouse_E18_Brain/data/SCEs/MouseEmbryo", "clustered.sceList.rds"
  ))
```

The cluster object is the most mature so also save it to permanent storage.

## Session Info

```{r}
sessionInfo()
```
