==> Mouse_E18_Brain/notebooks/MouseEmbryonicBrain/3A_GEX_marker_genes.Rmd <==
---
title: "GEX marker gene detection"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      smooth_scroll: true
    theme: readable
    highlight: tango 
    df_print: paged
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      dev = "png",
                      dpi = 300,
                      fig.asp=0.6, 
                      fig.path='./figures/cellplex-gex_markers/',
                      warning=FALSE, message=FALSE)
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = here::here())
suppressPackageStartupMessages({
  library(tidyverse)
  library(patchwork)
  library(scuttle)
  library(scater)
  library(scran)
  library(viridis)
  })

source(here::here("scripts/R/theme_publication.R"))
theme_set(theme_Publication())
```

# Aim

Perform marker gene annotation for downsstream maual annotation.

## Samples

Mouse embryo E18.5

# Load Data

From the 2B_QC_dimred notebook

```{r loadData}
sce_list <- readRDS(here::here(
  "Mouse_E18_Brain/data/SCEs/MouseEmbryo", "clustered.sceList.rds"
  ))

names(sce_list)

capture <- names(sce_list)
current <- sce_list[["CellPlex"]]
```

# Recap clusters in a UMAP {.tabset}

For this analysis we focus on the CellPlex dataset which has the highest quality cells.
I strongly suspect that cluster number 12 and 8 are low quality cells as the library size is low and the mitochondrial percent is high.  
Note all these cells have passed automated quality control filter based on 2 median absolute deviations.

## Cluster number

```{r umap_cluster}
plotReducedDim(current, dimred="UMAP", colour_by="cluster")
```

## Automated cell annotation broad cell type

```{r umap_broad_autoAnno}
plotReducedDim(current, dimred="UMAP", colour_by="broad_cellType") + ggtitle("")
```

## Automated cell annotation fine cell type

```{r umap_fine_autoAnno}
plotReducedDim(current, dimred="UMAP", colour_by="finecellType")
```

# CellPlex marker gene detection

CellPlex is the highest quality dataset. Use this one to find the marker genes.  
For each cluster, we rank candidate markers based on one of these effect size summaries.  
Use pairwise comparisons rather than comparing each cluster to the average of all other cells.

```{r}
marker.info <- scoreMarkers(current, groups=current$cluster)
```

Compare each pair of clusters and compute scores quantifying the differences in the expression distributions between clusters.  
The scores for all pairwise comparisons involving a particular cluster are then consolidated into a single DataFrame for that cluster.

## Iterate marker detection over each cluster

We choose the top 200 markers for each cluster to reduce the burden of visualisation.

```{r}
markers <- list()

for (i in 1:length(marker.info)) {
    chosen <- marker.info[[i]]
    ordered <- chosen[order(chosen$mean.AUC, decreasing=TRUE),]

    chosen <- chosen[order(chosen$mean.AUC, decreasing=TRUE),]
    markers[[i]] <- as.data.frame(chosen[c(1:200),])
}

names(markers) <- 1:length(markers)
```

Mung the list of marker genes.  
Convert to a summarized and usable form for all clusters

The p-values are said to be meaningless for this comparison. Quoted from the score markers documentation.  

*"the p-values for these types of comparisons are largely meaningless; individual cells are not meaningful units of experimental replication, while the groups themselves are defined from the data. Thus, by discarding the p-values, we can simplify our marker selection by focusing only on the effect sizes between groups."*

The meaning of the different columns in the score markers function:

* The logFC.cohen columns contain the standardized log-fold change, i.e., Cohen's d.
  + Difference in the mean log-expression for each group scaled by the average standard deviation across the two groups.
  + We can also interpret Cohen's d as the number of standard deviations between the two group means
* The AUC columns contain the area under the curve. This is the probability that a randomly chosen observation in one group is greater than a randomly chosen observation in the other group. 
  + Values greater than 0.5 indicate that a gene is upregulated in the first group.
* logFC.detected columns contain the log-fold change in the proportion of cells with detected (i.e., non-zero) expression between groups. 
  + This is specifically useful for detecting binary expression patterns, e.g., activation of an otherwise silent gene


```{r}
marker <- do.call(rbind, markers)
marker$gene <- row.names(marker)

marker <- as_tibble(marker) %>% 
  select(gene, mean.logFC.cohen, median.logFC.cohen, 
         mean.AUC, median.AUC, 
         mean.logFC.detected, median.logFC.detected) %>% 
  separate(gene, c("cluster", "gene.symbol"), sep = 2) %>% 
  filter(mean.logFC.cohen >= 0.5) %>% 
  filter(mean.AUC >= 0.5) %>% 
  filter(mean.logFC.detected >= 1)

# For some reason the gene symbol comes with a dot sometimes
# Its when cluster number is > 10
marker$gene.symbol <- gsub("\\.", "",marker$gene.symbol)
marker$cluster <- gsub("\\.", "",marker$cluster)

marker
```

## CellPlex visualise results

Take the top marker for each cluster into the dot plot

```{r}
topmarkers <- list()
for (i in (1:length(markers))) {
  topmarkers[[i]] <- markers[[i]][1,]
}

topmarkers <- do.call(rbind, topmarkers)
topmarkers <- row.names(topmarkers)

# subset for in expressed genes
keep_genes <- topmarkers %in% row.names(current)
topmarkers <- topmarkers[keep_genes]
```

### Dot plot

```{r cellplex_dot}
dot_unlabel <- plotDots(current, 
         features = topmarkers, 
         group="cluster", color=plasma(100)) +
  labs(title = "", y = "", x = "Cluster", color = "Log counts", size="Proportion of cells") +
  theme_Publication()

dot_unlabel
```

### Expression plots

```{r cellplex_expression, fig.dim=c(8,14)}
plotExpression(current, features=topmarkers, 
    x="cluster", colour_by="cluster", ncol=3) +
  theme_Publication()
```

# Christine's Biben requested markers

Hi Daniel,
 
I have been trying to annotate the brain dataset. Some clusters are easily identifiable but others are more difficult due to the fact that most of the genes if not all that other people use to recognize certain cell types are not listed on the DE lists. Would it be possible for me to access gene expression on the umap, or for you to send me some of the gene expression listed below, please?

```{r}
biben_marker <- c( 
"Tubb2b","Neurod2",
"Satb2","Fezf2","Olig1","Olig2",
"Pdgfra","Sox10","Apoe",
"Pdgfrb","Cspg4","Kcnj8","Reln",
"Lhx5","Egfr","Nrgn","Lpl",
"Slc17a6","Aldoc","Dlx1","Dlx2","Top2a","Ascl1",
"Sema3c","Neurod1","Gad1"
)
```

### Expression dot plot

```{r biben_dot, fig.dim=c(8,8)}
dot_unlabel <- plotDots(current, 
         features = biben_marker, 
         group="cluster", color=plasma(100)) +
  labs(title = "", y = "", x = "Cluster", color = "Log counts", size="Proportion of cells") +
  theme_Publication(base_size=12)

dot_unlabel
```

### UMAPs

Based on the dot plot I take the more informative genes to not crowd the UMAP

```{r}
umap_marker <- c( 
"Tubb2b","Neurod2",
"Satb2","Olig1",
"Apoe", "Dlx2",
"Pdgfrb","Kcnj8","Reln",
"Lhx5","Nrgn",
"Aldoc","Dlx1","Top2a",
"Sema3c","Neurod1"
)
```


```{r umap_biben, fig.dim=c(12,12)}
umap_plt <- list()

for (i in 1:length(umap_marker)) {
  umap_plt[[i]] <- plotReducedDim(current, dimred="UMAP", colour_by=umap_marker[i]) + 
    xlab("") + ylab("") +
    theme_Publication(base_size=10)
}

umap_plt[[1]] + umap_plt[[2]] + umap_plt[[3]] +  umap_plt[[4]] +
  umap_plt[[5]] + umap_plt[[6]] + umap_plt[[7]] +  umap_plt[[8]] +
  umap_plt[[9]] + umap_plt[[10]] + umap_plt[[11]] +  umap_plt[[12]] +
  umap_plt[[13]] + umap_plt[[14]] + umap_plt[[15]] +  umap_plt[[16]] +
     plot_layout(ncol = 4)
```

### Save markers

```{r}
write_csv(marker, here::here(
  "Mouse_E18_Brain/data/Matrices", "cellplex_top200-marker.csv"
))

# remove medians
marker <- marker %>% 
   select(cluster, gene.symbol, mean.logFC.cohen, mean.AUC, mean.logFC.detected)

write_csv(marker, here::here(
  "Mouse_E18_Brain/data/Matrices", "cellplex_simple_top200-marker.csv"
))

saveRDS(current,
  "/vast/scratch/users/brown.d/MouseEmbryo/SCEs/CellPlex_anno.rds"
  )
```

Write the counts as an aggreggated expression matrix.  
Chrisitine can manually check the counts if desired.

```{r}
summed <- aggregateAcrossCells(current, 
    id=colData(current)["cluster"])

summed <- counts(summed)
summed <- as.data.frame(summed[biben_marker,])
colnames(summed) <- seq(1:13)
write.csv(summed, "/vast/scratch/users/brown.d/MouseEmbryo/SCEs/marker_counts.csv")
```
## Session Info

```{r}
sessionInfo()
```
