==> Mouse_E18_Brain/notebooks/MouseEmbryonicBrain/3B_oligoTag_cluster_vis.Rmd <==
---
title: "Oligo tag Cluster Mouse brain"
author: "Daniel Brown"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      smooth_scroll: true
    theme: readable
    highlight: tango 
    df_print: paged
    code_folding: hide
editor_options: 
    chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      dev = "png",
                      dpi = 300,
                      fig.asp=0.6, 
                      fig.path='./figures/oligo_cluster_vis/',
                      warning=FALSE, message=FALSE)
knitr::opts_knit$set(root.dir = here::here())
suppressPackageStartupMessages({
  library(scater)
  library(scran)
  library(scuttle)
  library(patchwork)
  library(knitr)
  library(pheatmap)
  library(tidyverse)
})
source(here::here("scripts/R/theme_publication.R"))
theme_set(theme_Publication())
```

## Recap
The purpose of this experiment was to compare the performance of 3 different cell multiplexing reagents in Mouse Embryo E18.5  

## Aim

In this notebook the oligo tag counts are clustered and visualized in UMAP space.  

I base the analysis here on the PBMC notebook *oligo_QC_Norm.Rmd* I will not use all the code as the mouse embryo experiment is lower quality.

# Load Data

This list of SingleCellExperiment objects were generated in *GEX_QC_Norm* notebook.

```{r loadData}
sce_list <- readRDS(here::here(
  "Mouse_E18_Brain/data/SCEs/MouseEmbryo", "clustered.sceList.rds"
))
names(sce_list)

# Swap out the tag oligo slot for the gex slot as the primary count object
for (n in names(sce_list)) {
    current <- swapAltExp(sce_list[[n]], name="Multiplexing Capture", saved="GEX")
    sce_list[[n]] <- current
}
```

# Oligo Normalization

## Median normalization

Use the most basic normalisation as composition biases and other differences are of interest to my study. 

```{r normalise}
set.seed(666)

for (n in names(sce_list)) {
    current <- sce_list[[n]] 
    sf.geo <- geometricSizeFactors(current)
    sizeFactors(current) <- sf.geo
    current <- logNormCounts(current)
    sce_list[[n]] <- current
}

assayNames(current)
```

# Dimension reduction
We can apply downstream procedures like clustering and visualization on the log-normalized abundance matrix for the ADTs (Figure 12.7). Alternatively, if we had generated a matrix of PCs, we could use that as well.

```{r umap}
set.seed(999)

for (n in names(sce_list)) {
    current <- sce_list[[n]] 
    current <- runUMAP(current)
    sce_list[[n]] <- current
}
```

# Plot UMAP BFF CLuster call {.tabset}

BFF CLuster was the slightly better performing demultipelxing algorithm. Use this as the visualization colour.

## CellPlex 1:10 dilution

```{r umap_tagCluster_bff_CMO, fig.cap="CellPlex"}
p1 <- plotUMAP(sce_list[["CellPlex"]], 
         colour_by="bff_cluster") +
         ggtitle("CellPlex") + 
        theme(text = element_text(size = 18),
           legend.text = element_text(size = 14)) +
  theme_Publication()

p1
```

## MULTI-Seq LMO

```{r umap_tagCluster_bff_LMO, fig.cap="MULTI-Seq LMO"}
p2 <- plotUMAP(sce_list[["MULTI-Seq LMO"]], 
         colour_by="bff_cluster") +
         ggtitle("MULTI-Seq LMO") + 
        theme(text = element_text(size = 18),
           legend.text = element_text(size = 14)) +
  theme_Publication()

p2
```

## MULTI-Seq CMO

```{r umap_tagCluster_bff_NXT, fig.cap="MULTI-Seq CMO"}
p3 <- plotUMAP(sce_list[["MULTI-Seq CMO"]], 
         colour_by="bff_cluster") +
         ggtitle("MULTI-Seq CMO") + 
        theme(text = element_text(size = 18),
           legend.text = element_text(size = 14)) +
  theme_Publication()

p3
```

## UMAP together

```{r umap_combine, fig.asp=1, fig.cap="Combined UMAP"}
p1 + theme_Publication() +
  p2 + theme_Publication() + 
  p3 + theme_Publication() +
  plot_layout(guides = "collect", ncol = 2)

```

# Heatmaps {.tabset}

Heatmap is looking like the most useful visualization. Here I order the columns (cells) by the BFF Cluster calls instead of euclidean clustering of the oligo tag counts.

```{r sumcounts}
all_ave <- list()

for (n in names(sce_list)) {
    current <- sce_list[[n]]
    se.averaged <- sumCountsAcrossCells(current,
                                        ids=current$bff_cluster, # Change to donor identity when info is available
    exprs_values="logcounts", average=TRUE)
    all_ave[[n]] <- assay(se.averaged)
}

names(all_ave) <- names(sce_list)

fontSize=12
fontsizeRow = 16
legendFontSize = 6

# Specify colors
annotationColors = list(
    Demuxafy = c(unassigned="grey", singlet="skyblue", doublet="orange")
)

# Set colour scheme to sequential
col_fun <- colorRampPalette(c("#FFFFFF", "#FF0000"))(50)
```

## CellPlex

```{r cellplex_heat, fig.asp=0.95}
current <- "CellPlex"
set.seed(666)
cells <- sample(2500)

normcounts <- logcounts(sce_list[[current]][,cells])

# Update my_sample_col data when genotypes are available
my_sample_col <- as.data.frame(colData(
  sce_list[[current]][,cells]
  )[,c("bff_cluster", "Demuxafy_assignment")]
  )

my_sample_col <- my_sample_col[order(my_sample_col$bff_cluster),]
colnames(my_sample_col) <- c("BFF_Cluster", "Demuxafy")

pheatmap(normcounts[,row.names(my_sample_col)],
         annotation_col = my_sample_col,
         cluster_rows=FALSE,
         cluster_cols = FALSE,
         treeheight_col=0,
         show_rownames=TRUE,
         show_colnames = FALSE,
         annotation_colors=annotationColors,
         fontsize=fontSize,
         legend_font_size = legendFontSize,
         fontsize_row = fontsizeRow,
         main = current,
         color = col_fun,
         row.center = TRUE,
         labels_col=row.names(current)
         )
  
```

## MULTI-Seq LMO

```{r LMO_heat, fig.asp=0.95}
current <- "MULTI-Seq LMO"
set.seed(666)
cells <- sample(2500)

normcounts <- logcounts(sce_list[[current]][,cells])

# Update my_sample_col data when genotypes are available
my_sample_col <- as.data.frame(colData(
  sce_list[[current]][,cells]
  )[,c("bff_cluster", "Demuxafy_assignment")]
  )

my_sample_col <- my_sample_col[order(my_sample_col$bff_cluster),]
colnames(my_sample_col) <- c("BFF_Cluster", "Demuxafy")

pheatmap(normcounts[,row.names(my_sample_col)],
         annotation_col = my_sample_col,
         cluster_rows=FALSE,
         cluster_cols = FALSE,
         treeheight_col=0,
         show_rownames=TRUE,
         show_colnames = FALSE,
         annotation_colors=annotationColors,
         fontsize=fontSize,
         legend_font_size = legendFontSize,
         fontsize_row = fontsizeRow,
         main = current,
         color = col_fun,
         row.center = TRUE,
         labels_col=row.names(current)
         )
  
```

## MULTI-Seq CMO

```{r NXT_heat, fig.asp=0.95}
current <- "MULTI-Seq CMO"
set.seed(666)
cells <- sample(2500)

normcounts <- logcounts(sce_list[[current]][,cells])

# Update my_sample_col data when genotypes are available
my_sample_col <- as.data.frame(colData(
  sce_list[[current]][,cells]
  )[,c("bff_cluster", "Demuxafy_assignment")]
  )

my_sample_col <- my_sample_col[order(my_sample_col$bff_cluster),]
colnames(my_sample_col) <- c("BFF_Cluster", "Demuxafy")

pheatmap(normcounts[,row.names(my_sample_col)],
         annotation_col = my_sample_col,
         cluster_rows=FALSE,
         cluster_cols = FALSE,
         treeheight_col=0,
         show_rownames=TRUE,
         show_colnames = FALSE,
         annotation_colors=annotationColors,
         fontsize=fontSize,
         legend_font_size = legendFontSize,
         fontsize_row = fontsizeRow,
         main = current,
         color = col_fun,
         row.center = TRUE,
         labels_col=row.names(current)
         )
  
```

# Save SCE Objects

```{r saveSCEs, eval=FALSE}
for (n in names(sce_list)) {
    sce_list[[n]] <- swapAltExp(sce_list[[n]], "GEX")
}

saveRDS("/vast/scratch/users/brown.d/MouseEmbryo/SCEs/embryo.tag_dimred.rds")

```

## Session Info

```{r}
sessionInfo()
```
